// src/sdks/Newcomputation/computationV2ApiDocs.ts
export interface ComputationData {
  carbonEmission?: number;

  companyId?: number;

  computationId?: number;

  createBy?: number;

  createTime?: Date;

  dataStatus?: '0' | '1' | '2' | '3' | '4' | '5' | '6';

  dateIdx?: number;

  dateRange?: string;

  deleted?: boolean;

  id?: number;

  orgId?: number;

  orgName?: string;

  submitTime?: Date;

  updateBy?: number;

  updateTime?: Date;

  year?: number;
  dataStatus_name?: string;
}